pyproject.toml

[project]
name = "notifier"
requires-python = ">=3.13"
version = "0.1.0"


[tool.mypy]
check_untyped_defs = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
enable_error_code = [
    "deprecated",
    "explicit-override",
    "ignore-without-code",
    "mutable-override",
    "narrowed-type-not-subtype",
    "possibly-undefined",
    "redundant-expr",
    "redundant-self",
    "truthy-bool",
    "truthy-iterable",
    "unused-awaitable",
]
extra_checks = true
no_implicit_reexport = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.ruff]
indent-width = 4
line-length = 95
output-format = "concise"

[tool.ruff.format]
docstring-code-format = true
indent-style = "space"
line-ending = "lf"
quote-style = "double"


[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
mypy = "^1.16.1"
ruff = "^0.12.1"


[tool.poetry.dependencies]
py-cord = "^2.6.1"
python-dotenv = "^1.1.1"
audioop-lts = "^0.2.1"
